tests/ambient-light/AmbientLightSensor_basic.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: Basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var sensor, start_return, stop_return;
setup(function() {
  sensor = new AmbientLightSensor();
});

test(function() {
  assert_equals(sensor.state, "idle");
}, "The default sensor.state is 'idle'.");

test(function() {
  assert_equals(String(sensor.reading), "[object AmbientLightSensorReading]");
}, "the sensor.reading is AmbientLightSensorReading after executing start() method");

async_test(function(t) {
  sensor.onstatechange = t.step_func_done(function(event) {
    assert_equals(sensor.state, "activating");
  });
  sensor.onerror = t.step_func_done(function(event) {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  start_return = sensor.start();
}, "The sensor.state changes to 'activating' after sensor.start().");

//TODO: The permission is not ready.

test(function() {
  assert_equals(start_return, undefined);
}, "the sensor.start() return undefined");

test(function() {
  assert_throws("InvalidStateError", function() { sensor.start(); }, "start() twice");
}, "throw an InvalidStateError exception when state is neither idle nor errored");

async_test(function(t) {
  sensor.onstatechange = t.step_func_done(function(event) {
    assert_equals(sensor.state, "idle");
  });
  sensor.onerror = t.step_func_done(function(event) {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  stop_return = sensor.stop();
}, "The sensor.state changes to 'idle' after sensor.stop().");

test(function() {
  assert_equals(stop_return, undefined);
}, "the sensor.stop() returns undefined");

test(function() {
  assert_equals(String(sensor.reading), "null");
}, "the sensor.reading is null after executing stop() method");

test(function() {
  assert_throws("InvalidStateError", function() { sensor.stop(); }, "stop() twice");
}, "throw an InvalidStateError exception when state is either idle or errored");

</script>
